+++ proofs/src/plonk/trash/prover.rs
@@ -14,6 +14,7 @@ use crate::{
 #[cfg_attr(feature = "bench-internal", derive(Clone))]
 #[derive(Debug)]
 pub(crate) struct Committed<F: PrimeField> {
+    pub(crate) argument_index: usize,
     pub(crate) trash_poly: Polynomial<F, Coeff>,
 }
 
@@ -26,6 +27,7 @@ impl<F: WithSmallOrderMulGroup<3> + Ord> Argument<F> {
     #[allow(clippy::too_many_arguments)]
     pub(crate) fn commit<'a, 'params: 'a, CS, T>(
         &self,
+        argument_index: usize,
         params: &'params CS::Parameters,
         domain: &EvaluationDomain<F>,
         trash_challenge: F,
@@ -60,14 +62,17 @@ impl<F: WithSmallOrderMulGroup<3> + Ord> Argument<F> {
         let trash_commitment = CS::commit(
             params,
             &compressed_expression,
-            PolynomialLabel::Trash(self.name.clone()),
+            PolynomialLabel::Trash(argument_index),
         );
         let trash_poly = domain.lagrange_to_coeff(compressed_expression);
 
         // Hash permuted input commitment
         transcript.write(&trash_commitment)?;
 
-        Ok(Committed { trash_poly })
+        Ok(Committed {
+            argument_index,
+            trash_poly,
+        })
     }
 }
 
@@ -89,6 +94,11 @@ impl<F: WithSmallOrderMulGroup<3>> Committed<F> {
 
 impl<F: WithSmallOrderMulGroup<3>> Evaluated<F> {
     pub(crate) fn open(&self, x: F) -> impl Iterator<Item = ProverQuery<'_, F>> + Clone {
-        vec![ProverQuery::new(x, &self.committed.trash_poly)].into_iter()
+        vec![ProverQuery::new(
+            x,
+            &self.committed.trash_poly,
+            PolynomialLabel::Trash(self.committed.argument_index),
+        )]
+        .into_iter()
     }
 }