+++ proofs/src/plonk/trash/verifier.rs
@@ -12,6 +12,7 @@ use crate::{
 
 #[derive(Debug)]
 pub struct Committed<F: PrimeField, CS: PolynomialCommitmentScheme<F>> {
+    argument_index: usize,
     trash_commitment: CS::Commitment,
 }
 
@@ -24,6 +25,7 @@ pub struct Evaluated<F: PrimeField, CS: PolynomialCommitmentScheme<F>> {
 impl<F: PrimeField> Argument<F> {
     pub(crate) fn read_committed<CS: PolynomialCommitmentScheme<F>, T: Transcript>(
         &self,
+        argument_index: usize,
         transcript: &mut T,
     ) -> Result<Committed<F, CS>, Error>
     where
@@ -31,8 +33,11 @@ impl<F: PrimeField> Argument<F> {
     {
         let trash_commitment = transcript
             .read()
-            .map(|c: CS::Commitment| c.label(PolynomialLabel::Trash(self.name.clone())))?;
-        Ok(Committed { trash_commitment })
+            .map(|c: CS::Commitment| c.label(&[PolynomialLabel::Trash(argument_index)]))?;
+        Ok(Committed {
+            argument_index,
+            trash_commitment,
+        })
     }
 }
 
@@ -58,6 +63,7 @@ impl<F: WithSmallOrderMulGroup<3>, CS: PolynomialCommitmentScheme<F>> Evaluated<
         vec![VerifierQuery::new(
             x,
             &self.committed.trash_commitment,
+            PolynomialLabel::Trash(self.committed.argument_index),
             self.evaluated.trash_eval,
         )]
         .into_iter()